+++ includes/class-rest-volunteer.php
@@ -81,7 +81,7 @@ class Volunteer extends Base {
 						'validate_callback' => function ( $param ) {
 							return in_array(
 								$param,
-								[ 'orphan', 'address_fallback', 'missing_leeftijdsgroep', 'non_paying' ],
+								[ 'orphan', 'address_fallback', 'missing_leeftijdsgroep', 'former_members' ],
 								true
 							);
 						},
@@ -703,8 +703,8 @@ class Volunteer extends Base {
 			case 'missing_leeftijdsgroep':
 				$ids = $service->get_skipped_no_leeftijdsgroep_ids();
 				break;
-			case 'non_paying':
-				$ids = $service->get_non_paying_ids();
+			case 'former_members':
+				$ids = $service->get_former_member_ids();
 				break;
 			default:
 				return new \WP_Error( 'invalid_category', 'Unknown data-quality category.', [ 'status' => 400 ] );